just fixing some breakings

+++ FundRaisingServer/Models/DTOs/Case/CaseResponseDto.cs
@@ -4,7 +4,7 @@ namespace FundRaisingServer.Models.DTOs
     {
         public int CaseId { get; set; }
         public DateTime CreatedDate { get; set; }
-        public double TotalDonations { get; set; }
-        public double CollectedDonations { get; set; }
+        public decimal TotalDonations { get; set; }
+        public decimal CollectedDonations { get; set; }
     }
 }
